+++ internal/commands/queryrow.go
@@ -11,19 +11,20 @@ type QueryRow struct {
 	datasource *dbutils.DataSource
 }
 
-func (q *QueryRow) Init(context utils.Context, ds *dbutils.DataSource) {
-	q.datasource = ds
+func (c *QueryRow) Init(context utils.Context, ds *dbutils.DataSource) {
+	c.datasource = ds
 }
 
-func (q *QueryRow) Execute(args ...string) {
+func (c *QueryRow) Execute(args ...string) {
+	query := utils.OptionallyFromFile(args...)
 	s := make([]interface{}, len(args)-1)
 	for i := range s {
 		s[i] = args[i+1]
 	}
-	result, err := q.datasource.ExecuteQueryRow(args[0], s)
+	result, err := c.datasource.ExecuteQueryRow(query, s)
 	if err != nil {
-		log.Printf("ERROR: Database: %s, Query '%s' failed with error: %v\n", q.datasource.GetDBName(), args[0], err)
+		log.Printf("ERROR: Database: %s, Query '%s' failed with error: %v\n", c.datasource.GetDBName(), query, err)
 		return
 	}
-	fmt.Printf("Database: %s, Query '%s', result: %v\n", q.datasource.GetDBName(), args[0], result)
+	fmt.Printf("Database: %s, Query '%s', result: %v\n", c.datasource.GetDBName(), query, result)
 }