+++ internal/commands/snapshottable.go
@@ -11,15 +11,15 @@ type SnapshotTable struct {
 	datasource *dbutils.DataSource
 }
 
-func (s *SnapshotTable) Init(context utils.Context, ds *dbutils.DataSource) {
-	s.datasource = ds
+func (c *SnapshotTable) Init(context utils.Context, ds *dbutils.DataSource) {
+	c.datasource = ds
 }
 
-func (s *SnapshotTable) Execute(args ...string) {
+func (c *SnapshotTable) Execute(args ...string) {
 	query := fmt.Sprintf("INSERT INTO pgmaven_%s select * from %s;", args[0], args[0])
 
-	_, err := s.datasource.GetDatabase().Exec(query)
+	_, err := c.datasource.GetDatabase().Exec(query)
 	if err != nil {
-		log.Printf("ERROR: SnapShotTable insert failed with error: %s\n", err)
+		log.Printf("ERROR: Database '%s', SnapShotTable insert failed with error: %s\n", c.datasource.GetDBName(), err)
 	}
 }